Move to flat structure, simplify access points, add lookalikes set
