+++ README.md
@@ -13,13 +13,13 @@ Require a generator and pass a string from the dictionary
 const generate = require('nanoid/generate');
 const dictionary = require('nanoid-dictionary');
 
-const englishLowercase = generate(dictionary.alphabets.english.lowercase, 10);
+const englishLowercase = generate(dictionary.lowercase, 10);
 ```
 
 Or instead you can load only a particular dictionary
 
 ```javascript
-const englishLowercase = generate('nanoid-dictionary/alphabets/english/lowercase', 10);
+const englishLowercase = generate('nanoid-dictionary/lowercase', 10);
 ```
 
 
@@ -29,23 +29,45 @@ const englishLowercase = generate('nanoid-dictionary/alphabets/english/lowercase
 
 Numbers from 0 to 9
 
+```javascript
+const numbers = require('nanoid-dictionary/numbers');
+```
+
 ### `alphabets`
 
-#### `english`
+#### `lowercase`
 
-##### `lowercase`
+Lowercase English letters.
 
-Lowercase English letters
+Available both as `alphabets/lowercase` and simply `lowercase`.
 
-##### `uppercase`
+```javascript
+const englishLowercase = require('nanoid-dictionary/lowercase');
+```
 
-Uppercase English letters
+#### `uppercase`
+
+Uppercase English letters.
+
+Available both as `alphabets/uppercase` and simply `uppercase`.
+
+```javascript
+const englishUppercase = require('nanoid-dictionary/uppercase');
+```
+
+### `lookalikes`
+
+Characters that have lookalikes: `1`, `l`, `0`, `O`
+
+```javascript
+const lookalikes = require('nanoid-dictionary/lookalikes');
+```
 
 
 
 ## Helper functions
 
-### `preventMisreadings(str, chars)`
+### `preventMisreadings(str, unsafeChars)`
 
 It accepts a string and removes all the characters that look similar. You can pass your own optional character set if you want.
 
@@ -60,18 +82,17 @@ const safeString = preventMisreadings(unsafeString); \\ 'ab'
 
 A string to process.
 
-#### `chars`
+#### `unsafeChars`
 
-An optional array of unsafe characters.
+An optional string that contains only unsafe characters.
 
-Deault unsafe characters are: `1`, `l`, `0`, `o`
-
+Deault unsafe character set is taken from `lookalikes`.
 
 ```javascript
 
 const preventMisreadings = require('nanoid-dictionary/preventMisreadings');
 const unsafeString = 'a`\'b';
-const unsafeChars = ['`', '\''];
+const unsafeChars = '`\'';
 
 const safeString = preventMisreadings(unsafeString, unsafeChars); \\ 'ab'
