+++ filename.js
@@ -1,0 +1,5 @@
+var lowercase = require('./lowercase');
+var uppercase = require('./uppercase');
+var numbers = require('./numbers');
+
+module.exports = '_-,' + numbers + lowercase + uppercase;
