--- legacy/filename.js
@@ -1,5 +1,0 @@
-var lowercase = require('../alphabets/english/lowercase');
-var uppercase = require('../alphabets/english/uppercase');
-var numbers = require('../numbers');
-
-module.exports = '_-,' + numbers + alphabets.english.lowercase + alphabets.english.uppercase;
