+++ lookalikes.js
@@ -1,0 +1,1 @@
+module.exports = '1l0o';
