+++ preventMisreadings.js
@@ -1,11 +1,12 @@
-var confusingChars = ['1', 'l', '0', 'o'];
+var lookalikes = require('./lookalikes');
+
 var matchAllOccurences = function(symbol) {
     return new RegExp(symbol + '+', 'gi');
 };
 
 module.exports = function(str, chars) {
-    chars = chars || confusingChars;
-    chars.forEach(function(char) {
+    chars = chars || lookalikes;
+    chars.split('').forEach(function(char) {
         str = str.replace(matchAllOccurences(char), '');
     })
     return str;
