+++ uppercase.js
